requirements.txt
discord.py==2.3.2
spotipy==2.24.0
groq==0.9.0
python-dotenv==1.0.1
pytest==8.3.2
pytest-asyncio==0.24.0
